fix: fix filtered tracks hook
