+++ src/components/AudioPlayer/PlayerVolume.tsx
@@ -1,6 +1,6 @@
 import { PlayerBtn } from './PlayerBtn';
-import { forwardRef, useState } from 'react';
 import svgIcon from '../../assets/svg/sprite.svg';
+import { forwardRef, useEffect, useState } from 'react';
 
 import styles from './styles.module.css';
 
@@ -10,18 +10,20 @@ type PlayerVolumeProps = {
 
 export const PlayerVolume = forwardRef<HTMLAudioElement, PlayerVolumeProps>(
   ({ onClick }, ref) => {
-    const [volume, setVolume] = useState(100);
+    const [volume, setVolume] = useState(50);
 
     const changeVolumeHandler = (
       event: React.ChangeEvent<HTMLInputElement>
     ) => {
       const { value } = event.target;
       setVolume(Number(value));
+    };
 
+    useEffect(() => {
       if (ref !== null && typeof ref !== 'function') {
         ref.current!.volume = volume / 100;
       }
-    };
+    }, [ref, volume]);
 
     return (
       <div className={styles.volume}>
