+++ src/components/Playlist/styles.module.css
@@ -2,43 +2,17 @@
   margin-top: 36px;
 }
 
-.header {
+.tracks {
   display: flex;
-  align-items: center;
-  justify-content: space-between;
+  flex-direction: column;
+  gap: 12px;
 
-  margin-bottom: 12px;
-
-  font-size: 14px;
-  line-height: 24px;
-  letter-spacing: 2px;
-
-  color: #696969;
+  max-height: 100vh;
+  overflow-y: scroll;
+  scrollbar-width: none;
 }
 
-.track {
-  width: 42%;
-}
-
-.performer {
-  width: 29%;
-}
-
-.album {
-  width: 22%;
-}
-
-.time {
-  display: flex;
-  flex-grow: 1;
-  justify-content: flex-end;
-}
-
-.timeSvg {
-  margin: 0;
-  width: 12px;
-  height: 12px;
-
-  fill: transparent;
-  stroke: #696969;
+.tracks::-webkit-scrollbar {
+  width: 0px;
+  background-color: transparent;
 }
