+++ src/pages/CollectionPage.tsx
@@ -2,9 +2,9 @@ import { useParams } from 'react-router-dom';
 import { Title } from '../components/Title/Title';
 import { Search } from '../components/Search/Search';
 import { Sidebar } from '../components/Sidebar/Sidebar';
-import { useGetCollectionQuery } from '../store/api/tracksAPI';
 import { Playlist } from '../components/Playlist/Playlist';
 import { useFilterTracks } from '../hooks/useFilteredTracks';
+import { useGetCollectionQuery } from '../store/api/tracksAPI';
 import { Navigation } from '../components/Navigation/Navigation';
 import { AudioPlayer } from '../components/AudioPlayer/AudioPlayer';
 
@@ -22,7 +22,7 @@ export const CollectionPage = () => {
         <div className="main-content">
           <Search query={searchQuery} onChange={changeHandler} />
           <Title text={data?.name} />
-          <Playlist tracks={filteredTracks} loading={isLoading} />
+          <Playlist tracks={filteredTracks} isLoading={isLoading} />
         </div>
         <Sidebar isOffersShown={false} />
       </main>
